refactor: displayMenu, readInput, InputError case
